src/test/java/org/jasig/portlet/cms/service/dao/AwsS3BucketContentDaoTest.java
/**
 * Licensed to Apereo under one or more contributor license
 * agreements. See the NOTICE file distributed with this work
 * for additional information regarding copyright ownership.
 * Apereo licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file
 * except in compliance with the License.  You may obtain a
 * copy of the License at the following location:
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package org.jasig.portlet.cms.service.dao;

import static org.mockito.BDDMockito.given;
import static org.mockito.BDDMockito.willThrow;
import static org.mockito.Matchers.any;
import static org.mockito.Mockito.verify;

import java.util.ArrayList;
import java.util.List;

import javax.portlet.ActionRequest;
import javax.portlet.PortletPreferences;
import javax.portlet.PortletRequest;

import org.apache.commons.lang.NullArgumentException;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.amazonaws.AmazonClientException;
import com.amazonaws.AmazonServiceException;
import com.amazonaws.services.s3.AmazonS3;
import com.amazonaws.services.s3.model.GetObjectRequest;
import com.amazonaws.services.s3.model.PutObjectRequest;
import com.amazonaws.services.s3.model.S3Object;
import com.amazonaws.services.s3.model.S3ObjectInputStream;

/**
 * JUnit test class for {@link AwsS3BucketContentDao}.
 */
public class AwsS3BucketContentDaoTest {

    @Mock private AmazonS3 amazonS3Client;
    @Mock private ContentAwsS3BucketConfig awsS3BucketConfig;
    @Mock private PortletPreferences portletPreferences;
    @Mock private ActionRequest actionRequest;
    @Mock private PortletRequest portletRequest;

    @Mock private S3Object s3Object;
    @Mock private S3ObjectInputStream s3ObjectInputStream;

    private String content;
    private String localeKey;

    private AwsS3BucketContentDao dao;

    @Before
    public void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);
        this.content = "<h1>BLAH, BLAH, BLAH</h1>";
        this.localeKey = "";
        given(this.s3Object.getObjectContent()).willReturn(this.s3ObjectInputStream);
        given(this.actionRequest.getPreferences()).willReturn(this.portletPreferences);
        given(this.portletRequest.getPreferences()).willReturn(this.portletPreferences);
    }

    @After
    public void tearDown() throws Exception {
    }

    @Test (expected = NullArgumentException.class)
    public void constructorThrowsExceptionIfClientIsNull() {
        // when
        new AwsS3BucketContentDao(null, this.awsS3BucketConfig);
    }

    @Test (expected = NullArgumentException.class)
    public void constructorThrowsExceptionIfConfigIsNull() {
        // when
        new AwsS3BucketContentDao(this.amazonS3Client, null);
    }

    @Test (expected = AwsS3BucketContentDao.TokenValueNotFoundException.class)
    public void getContentMethodThrowsExceptionIfObjectFormatTokenValueNotFoundInPortletPrefs() {
        // given
        final String tokenKey = "abcId";
        final List<String> tokenKeyList = this.createTokenKeysList(tokenKey);
        this.objectKeyFormatHasTokensConfig(tokenKeyList);
        // when
        this.dao.getContent(this.portletRequest	, this.localeKey);
    }

// Getting NPE for this one in code where input stream is read; need to address this to fix test.	
//    @Test
//    public void getContentMethodDoesNotThrowTokenValueNotFoundExceptionIfObjectFormatTokenValuesFoundInPortletPrefs() {
//        // given
//        final String tokenKey = "abcId";
//        final List<String> tokenKeyList = this.createTokenKeysList(tokenKey);
//        this.objectKeyFormatHasTokensConfig(tokenKeyList);
//        given(this.portletPreferences.getValue(tokenKey, null)).willReturn("123");
//        // when
//        this.dao.getContent(this.portletRequest	, this.localeKey);
//        // then
//        verify(this.amazonS3Client).getObject(any(GetObjectRequest.class));
//    }

    @Test (expected = ContentPersistenceException.class)
    public void getContentMethodTranslatesCaughtAmazonClientException() {
        // given
        this.happyPathConfig();
        this.getObjectRequestThrowsAmazonClientException();
        // when
        this.dao.getContent(this.portletRequest, this.localeKey);
    }

    @Test (expected = ContentPersistenceException.class)
    public void getContentMethodTranslatesCaughtAmazonServiceException() {
        // given
        this.happyPathConfig();
        this.getObjectRequestThrowsAmazonServiceException();
        // when
        this.dao.getContent(this.portletRequest, this.localeKey);
    }

    @Test (expected = AwsS3BucketContentDao.TokenValueNotFoundException.class)
    public void saveContentMethodThrowsExceptionIfObjectFormatTokenValueNotFoundInPortletPrefs() {
        // given
        final String tokenKey = "abcId";
        final List<String> tokenKeyList = this.createTokenKeysList(tokenKey);
        this.objectKeyFormatHasTokensConfig(tokenKeyList);
        // when
        this.dao.saveContent(this.actionRequest, this.content, this.localeKey);
    }

    @Test
    public void saveContentMethodDoesNotThrowTokenValueNotFoundExceptionIfObjectFormatTokenValuesFoundInPortletPrefs() {
        // given
        final String tokenKey = "abcId";
        final List<String> tokenKeyList = this.createTokenKeysList(tokenKey);
        this.objectKeyFormatHasTokensConfig(tokenKeyList);
        given(this.portletPreferences.getValue(tokenKey, null)).willReturn("123");
        // when
        this.dao.saveContent(this.actionRequest, this.content, this.localeKey);
        // then
        verify(this.amazonS3Client).putObject(any(PutObjectRequest.class));
    }

    @Test (expected = ContentPersistenceException.class)
    public void putContentMethodTranslatesCaughtAmazonClientException() {
        // given
        this.happyPathConfig();
        this.putObjectRequestThrowsAmazonClientException();
        // when
        this.dao.saveContent(this.actionRequest, this.content, this.localeKey);
    }

    @Test (expected = ContentPersistenceException.class)
    public void putContentMethodTranslatesCaughtAmazonServiceException() {
        // given
        this.happyPathConfig();
        this.putObjectRequestThrowsAmazonServiceException();
        // when
        this.dao.saveContent(this.actionRequest, this.content, this.localeKey);
    }

    private void happyPathConfig() {
        given(this.awsS3BucketConfig.getBucketName()).willReturn("mybucket-0138383");
        given(this.awsS3BucketConfig.getObjectKeyFormat()).willReturn("simple/format/blah.html");
        this.dao = new AwsS3BucketContentDao(this.amazonS3Client, this.awsS3BucketConfig);
    }

    private void objectKeyFormatHasTokensConfig(final List<String> tokenKeysList) {
        given(this.awsS3BucketConfig.getBucketName()).willReturn("mybucket-0138383");
        given(this.awsS3BucketConfig.getObjectKeyFormat()).willReturn(this.createTestObjectKeyFormat(tokenKeysList));
        given(this.awsS3BucketConfig.getObjectKeyFormatTokensPortletPrefKeysList()).willReturn(tokenKeysList);
        this.dao = new AwsS3BucketContentDao(this.amazonS3Client, this.awsS3BucketConfig);
    }

    private List<String> createTokenKeysList(final String... tokenKeys) {
        final List<String> result = new ArrayList<String>(tokenKeys.length);
        for (String key : tokenKeys) {
            result.add(key);
        }
        return result;
    }

    private String createTestObjectKeyFormat(final List<String> tokenKeys) {
        String result = "simple/";
        for (String key : tokenKeys) {
            final String token = String.format(AwsS3BucketContentDao.OBJECT_KEY_FORMAT_PORTLET_PREF_TOKEN_FORMAT, key);
            result += token + "/";
        }
        result += "blah.html";
        return result;
    }

    private void getObjectRequestThrowsAmazonClientException() {
        willThrow(new AmazonClientException("")).given(this.amazonS3Client).getObject(any(GetObjectRequest.class));
    }
    private void getObjectRequestThrowsAmazonServiceException() {
        willThrow(new AmazonServiceException("")).given(this.amazonS3Client).getObject(any(GetObjectRequest.class));
    }
    private void putObjectRequestThrowsAmazonClientException() {
        willThrow(new AmazonClientException("")).given(this.amazonS3Client).putObject(any(PutObjectRequest.class));
    }
    private void putObjectRequestThrowsAmazonServiceException() {
        willThrow(new AmazonServiceException("")).given(this.amazonS3Client).putObject(any(PutObjectRequest.class));
    }

}
